Split remaining gate-related header files into header/source
